feat: notificationSendService test
