+++ src/main/java/com/LetMeDoWith/LetMeDoWith/presentation/notification/controller/NotificationController.java
@@ -1,0 +1,2 @@
+package com.LetMeDoWith.LetMeDoWith.presentation.notification.controller;public class NotificationController {
+}
