+++ src/main/resources/db/migration/V0.4.0__create_notification_tables.sql
