+++ src/test/java/com/LetMeDoWith/LetMeDoWith/application/notification/NotificationSendServiceTest.java
@@ -1,0 +1,2 @@
+package com.LetMeDoWith.LetMeDoWith.application.notification;public class NotificationSendServiceTest {
+}
